BREAKING_CHANGES(MTMC): switched SimulationConfig.thread_count from `ushort` to `int` for ease of use

+++ src/MTMCNET/SimulationConfig.cs
@@ -9,7 +9,7 @@ public class SimulationConfig<T>
   public Func<IRandom, T> sim_obj_ctor   = null!;
   public ulong target_iteration          = 1_000_000_000;
   public float check_rate                = 0.01f;
-  public ushort thread_count             = (ushort)(Environment.ProcessorCount / 2);
+  public int thread_count                = Environment.ProcessorCount / 2;
   public List<Func<IRandom>> rng_ctor    = null!;
   public bool periodic_check_prints_body = false;
   public TimeSpan minimum_wait           = TimeSpan.FromMilliseconds(2368);
